fix(func): use reference semantic (#62)
`Func` is now a reference object. Solves #56 

There are two main reasons for this:

1. The order of arguments in a function is in the same order in which they are created with `hlo_input()`.
2. Functions that have no causal impact on the return values are still added to the stablehlo program.
    This will e.g. be important for custom calls, that just print, but don't return anything.
    
The only thing that really changes from a usability perspective is that now, a function needs to be created before starting to define inputs.

```r
func <- hlo_func("myfunc")
x <- hlo_input("x", "f32", c(2, 2), func = func)
```

We can also leave out the `func = func`, because the default is to use `.current_func()`, which is the most recently generated function.